채용 지원 미리보기 작업

+++ src/main/java/com/kb/inno/admin/DTO/RecruitDTO.java
@@ -19,7 +19,7 @@ public class RecruitDTO {
     private int employ_sn;       // 채용 일련번호
     private String employ_yr;    // 연도
     private int employ_emp_nocs; // 채용 직원 수
-    private int sprt_amt;        // 지원 금액
+    private double sprt_amt;        // 지원 금액
     private int employ_ent_nocs; // 채용 기업 수
     private int frst_rgtr;       // 최초 등록자
     private Date frst_reg_dt;    // 최초 등록 일시
